Move LazyColumn to PodcastContent component and add "Scroll to episodes" behavior
